create chroma vector db

+++ airbyte-integrations/connectors/destination-chroma/destination_chroma/indexer.py
@@ -14,31 +14,39 @@ from airbyte_cdk.models import ConfiguredAirbyteCatalog
 from airbyte_cdk.models.airbyte_protocol import DestinationSyncMode
 
 from destination_chroma.config import ChromaIndexingConfigModel
-from .utils import convert_to_valid_collection_name
+from .utils import validate_collection_name
 
 
 class ChromaIndexer(Indexer):
     def __init__(self, config: ChromaIndexingConfigModel, embedder: Embedder):
         super().__init__(config, embedder)
         self.client = self._get_client()
+        self.collection_name = validate_collection_name(config.collection_name)
 
     def check(self):
         try:
-            client = self.client
-            client.heartbeat()
+            heartbeat = self.client.heartbeat()
+            if not heartbeat:
+                return "Chroma client server is not alive"
+            collection = self.client.get_or_create_collection(self.collection_name)
+            count = collection.count()
+            if count!=0 and not count:
+                return f"unable to get or create collection with name {self.collection_name}"
             return
         except Exception as e:
             return format_exception(e)
 
     def index(self, document_chunks: List[Chunk], delete_ids: List[str]) -> None:
         if len(delete_ids) > 0:
-            self._delete_records(METADATA_RECORD_ID_FIELD, delete_ids)
+            self._delete_records(delete_ids=delete_ids)
         embedding_vectors = self.embedder.embed_texts([chunk.page_content for chunk in document_chunks])
         entities = []
-        # TODO separate text, metadata and vectors ----- also get the stream name before writing
         for i in range(len(document_chunks)):
             chunk = document_chunks[i]
-            entities.append({**chunk.metadata, self.config.vector_field: embedding_vectors[i], self.config.text_field: chunk.page_content})
+            entities.append({'id': chunk.metadata[METADATA_RECORD_ID_FIELD], 
+                             'embedding': embedding_vectors[i], 
+                             'metadata': chunk.metadata,
+                             'document': chunk.page_content})
         self._write_data(entities)
 
     def pre_sync(self, catalog: ConfiguredAirbyteCatalog) -> None:
@@ -46,8 +54,7 @@ class ChromaIndexer(Indexer):
         streams_to_overwrite = [stream 
                                 for stream in catalog.streams 
                                 if stream.destination_sync_mode == DestinationSyncMode.overwrite]
-        # TODO delete for each stream
-        self._delete_records(METADATA_STREAM_FIELD, streams_to_overwrite)
+        self._delete_records(field_name=METADATA_STREAM_FIELD, field_values=streams_to_overwrite)
 
     @staticmethod
     def _get_client(self):
@@ -71,26 +78,28 @@ class ChromaIndexer(Indexer):
             return client
         return
     
-    def _delete_records(self, field_name, field_values):
-        where_dict_list = [{field_name:value} for value in field_values]
-        collection_name: str
-        collection = self.client.get_collection(collection_name)
+    def _delete_records(self, field_name=None, field_values=None, delete_ids=None):
+        collection = self.client.get_collection(self.collection_name)
         try:
-            for where_dict in where_dict_list:
-                collection.delete(where=where_dict)
+            if delete_ids:
+                collection.delete(ids=delete_ids)
+            if field_name and field_values:
+                where_dict_list = [{field_name:value} for value in field_values]
+                for where_dict in where_dict_list:
+                    collection.delete(where=where_dict)
         except ValueError as e:
             raise e
 
     def _write_data(self, entities):
-        record_id = self.get_record_id(record=record)
-        embedding = self.get_embedding(record=record)
-        record = self.prepare_record(record=record)
-        collection_name = convert_to_valid_collection_name()
+        ids = [entity['id'] for entity in entities]
+        embeddings = [entity['embedding'] for entity in entities]
+        metadata = [entity['metadata'] for entity in entities]
+        documents = [entity['document'] for entity in entities]
 
-        collection = self.client.get_collection(name=collection_name)
+        collection = self.client.get_collection(name=self.collection_name)
         collection.add(
-            embeddings=embedding,
-            documents=[record], # TODO get specific records to add
-            ids=[record_id],
-            metadata=None # TODO add support for metadata
+            ids=ids,
+            embeddings=embeddings,
+            metadata=metadata,
+            documents=documents
         )
